Hw14 (#15)
* ДЗ 14

* ДЗ 14: fix encoding

+++ payment_detail_api_pack.pkb
@@ -1,23 +1,13 @@
 ﻿create or replace package body payment_detail_api_pack
 is
 
-procedure check_null
-    (p_param in number,
-    p_param_name in varchar2)
-is
-begin
-    if p_param is null then
-        dbms_output.put_line (p_param_name || ': ' || c_message_error_field_is_null);
-    end if;
-end;
-
 procedure print_result
     (p_payment_id in payment.payment_id%type,
     p_message in varchar2,
     p_message2 in varchar2 default null)
 is
 begin
-    dbms_output.put_line ('(' || to_char (systimestamp, c_timestamp_format) || ') ID платежа: ' || to_char (p_payment_id) || ' - ' || p_message || p_message2);
+    dbms_output.put_line ('(' || to_char (systimestamp, payment_constant_pack.c_timestamp_format) || ') ID платежа: ' || to_char (p_payment_id) || ' - ' || p_message || p_message2);
 end print_result;
 
 procedure insert_or_update_payment_detail
@@ -25,18 +15,20 @@ procedure insert_or_update_payment_detail
     p_payment_detail in t_payment_detail_array)
 is
 begin
-    check_null (p_payment_id, 'p_payment_id');
+    if p_payment_id is null then
+        raise payment_constant_pack.e_invalid_input_parameter;
+    end if;
 
     if p_payment_detail is null or not p_payment_detail is not empty then
-        dbms_output.put_line (c_message_error_not_init_or_empty);
+        raise payment_constant_pack.e_invalid_collection;
     else
         for i in p_payment_detail.first .. p_payment_detail.last
         loop
             if p_payment_detail(i).field_id is null then
-                dbms_output.put_line (c_message_error_field_id_is_null);
+                raise payment_constant_pack.e_invalid_collection_field_id;
             end if;
             if p_payment_detail(i).field_value is null then
-                dbms_output.put_line (c_message_error_field_value_is_null);
+                raise payment_constant_pack.e_invalid_collection_field_value;
             end if;
             dbms_output.put_line ('Field_id: ' || p_payment_detail(i).field_id || '. Field_value: ' || p_payment_detail(i).field_value);
         end loop;
@@ -65,7 +57,18 @@ begin
             t.field_id,
             t.field_value);
 
-    print_result (p_payment_id, c_message_insert_or_update);
+    print_result (p_payment_id, payment_constant_pack.c_message_insert_or_update);
+exception
+    when payment_constant_pack.e_invalid_input_parameter then
+        raise_application_error (payment_constant_pack.e_invalid_input_parameter_code, payment_constant_pack.e_invalid_input_parameter_message);
+    when payment_constant_pack.e_invalid_collection then
+        raise_application_error (payment_constant_pack.e_invalid_collection_code, payment_constant_pack.e_invalid_collection_message);
+    when payment_constant_pack.e_invalid_collection_field_id then
+        raise_application_error (payment_constant_pack.e_invalid_collection_field_id_code, payment_constant_pack.e_invalid_collection_field_id_message);
+    when payment_constant_pack.e_invalid_collection_field_value then
+        raise_application_error (payment_constant_pack.e_invalid_collection_field_value_code, payment_constant_pack.e_invalid_collection_field_value_message);
+    when others then
+        raise_application_error (payment_constant_pack.e_other_code, payment_constant_pack.e_other_message);
 end insert_or_update_payment_detail;
 
 procedure delete_payment_detail
@@ -73,10 +76,12 @@ procedure delete_payment_detail
     p_payment_detail_field_ids in t_number_array)
 is
 begin
-    check_null (p_payment_id, 'p_payment_id');
+    if p_payment_id is null then
+        raise payment_constant_pack.e_invalid_input_parameter;
+    end if;
 
     if p_payment_detail_field_ids is null or not p_payment_detail_field_ids is not empty then
-        dbms_output.put_line (c_message_error_not_init_or_empty);
+        raise payment_constant_pack.e_invalid_collection;
     end if;
 
     delete from payment_detail pd
@@ -86,7 +91,14 @@ begin
             (select t.column_value as field_id
             from table (p_payment_detail_field_ids) t);
 
-    print_result (p_payment_id, c_message_delete, ' Количество удаляемых полей: ' || to_char (p_payment_detail_field_ids.count()));
+    print_result (p_payment_id, payment_constant_pack.c_message_delete, ' Количество удаляемых полей: ' || to_char (p_payment_detail_field_ids.count()));
+exception
+    when payment_constant_pack.e_invalid_input_parameter then
+        raise_application_error (payment_constant_pack.e_invalid_input_parameter_code, payment_constant_pack.e_invalid_input_parameter_message);
+    when payment_constant_pack.e_invalid_collection then
+        raise_application_error (payment_constant_pack.e_invalid_collection_code, payment_constant_pack.e_invalid_collection_message);
+    when others then
+        raise_application_error (payment_constant_pack.e_other_code, payment_constant_pack.e_other_message);
 end delete_payment_detail;
 
 end payment_detail_api_pack;
